Changed to double precision to ensure best possible comparison between methods

+++ analyze.C
@@ -46,11 +46,11 @@ TBranch *b_v2;
 TBranch *b_pt;
 TBranch *b_phi;
 
-float calc4_event(float, float, float, float, float);
-float calc4_track(float, float, float, float, float, float, float, float, float);
+double calc4_event(double, double, double, double, double);
+double calc4_track(double, double, double, double, double, double, double, double, double);
 
-float calc4_event_YZ(float, float, float, float, float);
-float calc4_track_YZ(float, float, float, float, float, float, float, float, float);
+double calc4_event_YZ(double, double, double, double, double);
+double calc4_track_YZ(double, double, double, double, double, double, double, double, double);
 
 
 // --- from generic forumulas ----------------------------------------------------
@@ -78,7 +78,7 @@ int main(int argc, char *argv[])
   char outFile[100];
 
   char c_offset[100];
-  float offset = 0;
+  double offset = 0;
 
   if(argc==1)
     {
@@ -235,7 +235,7 @@ int main(int argc, char *argv[])
       int mult = d_mult;
       th1d_mult->Fill(mult);
 
-      float psi2true = d_psi2;
+      double psi2true = d_psi2;
       th1d_psi2->Fill(psi2true);
 
       // --- for the generic formulas ---------
@@ -248,32 +248,32 @@ int main(int argc, char *argv[])
         } // for(Int_t h=0;h<maxHarmonic;h++)
       // --------------------------------------
       // --- first track loop, q-vectors
-      float Q2x = 0;
-      float Q2y = 0;
-      float Q4x = 0;
-      float Q4y = 0;
-      float Q2x_sub = 0;
-      float Q2y_sub = 0;
-      float Q4x_sub = 0;
-      float Q4y_sub = 0;
-      float Q2xprime = 0;
-      float Q2yprime = 0;
-      float Q4xprime = 0;
-      float Q4yprime = 0;
-      float Q2x_subprime = 0;
-      float Q2y_subprime = 0;
-      float Q4x_subprime = 0;
-      float Q4y_subprime = 0;
+      double Q2x = 0;
+      double Q2y = 0;
+      double Q4x = 0;
+      double Q4y = 0;
+      double Q2x_sub = 0;
+      double Q2y_sub = 0;
+      double Q4x_sub = 0;
+      double Q4y_sub = 0;
+      double Q2xprime = 0;
+      double Q2yprime = 0;
+      double Q4xprime = 0;
+      double Q4yprime = 0;
+      double Q2x_subprime = 0;
+      double Q2y_subprime = 0;
+      double Q4x_subprime = 0;
+      double Q4y_subprime = 0;
       for(int itrk=0; itrk<mult; itrk++)
 	{
-	  float phi = d_phi[itrk];
-	  float x  = cos(phi);
-	  float y  = sin(phi);
-	  float xoff = offset;
-	  float yoff = offset;
-	  float xprime = x - xoff;
-	  float yprime = y - yoff;
-	  float phiprime = atan2(yprime,xprime);
+	  double phi = d_phi[itrk];
+	  double x  = cos(phi);
+	  double y  = sin(phi);
+	  double xoff = offset;
+	  double yoff = offset;
+	  double xprime = x - xoff;
+	  double yprime = y - yoff;
+	  double phiprime = atan2(yprime,xprime);
           // --- from generic formulas ----------------------------------------------------------------------
           Double_t dPhi = 0.0; // particle angle
           Double_t wPhi = 1.0; // particle weight
@@ -306,18 +306,18 @@ int main(int argc, char *argv[])
 	  Q4x_subprime += cos(4*phiprime);
 	  Q4y_subprime += sin(4*phiprime);
 	} // End of track loop
-      //float psi2reco = atan2(Q2y,Q2x);
-      float psi2reco = atan2(Q2y_sub,Q2x_sub)/2;
+      //double psi2reco = atan2(Q2y,Q2x);
+      double psi2reco = atan2(Q2y_sub,Q2x_sub)/2;
       th1d_psi2_reco->Fill(psi2reco);
       th1d_psi2_tmr->Fill(cos(2*(psi2true-psi2reco)));
       tp1d_psi2_tmr->Fill(1,cos(2*(psi2true-psi2reco)));
-      float psi2primereco = atan2(Q2y_subprime,Q2x_subprime)/2;
+      double psi2primereco = atan2(Q2y_subprime,Q2x_subprime)/2;
       th1d_psi2prime_reco->Fill(psi2primereco);
       th1d_psi2prime_tmr->Fill(cos(2*(psi2true-psi2primereco)));
       tp1d_psi2prime_tmr->Fill(1,cos(2*(psi2true-psi2primereco)));
-      float two = ( Q2x*Q2x + Q2y*Q2y - mult) / (mult*mult - mult);
-      float four = calc4_event(Q2x,Q2y,Q4x,Q4y,mult);
-      float fourYZ = calc4_event_YZ(Q2x,Q2y,Q4x,Q4y,mult);
+      double two = ( Q2x*Q2x + Q2y*Q2y - mult) / (mult*mult - mult);
+      double four = calc4_event(Q2x,Q2y,Q4x,Q4y,mult);
+      double fourYZ = calc4_event_YZ(Q2x,Q2y,Q4x,Q4y,mult);
 
       tp1d_c2->Fill(1,two);
       tp1d_c2mult->Fill(mult,two);
@@ -356,51 +356,51 @@ int main(int argc, char *argv[])
       for(int itrk=0; itrk<mult; itrk++)
 	{
 	  int charge = d_charge[itrk];
-	  float phi = d_phi[itrk];
-	  float pt = d_pt[itrk];
-	  float v2 = d_v2[itrk];
+	  double phi = d_phi[itrk];
+	  double pt = d_pt[itrk];
+	  double v2 = d_v2[itrk];
 	  th1d_phi->Fill(phi);
 	  th1d_pt->Fill(pt);
-	  float v2_track_true = cos(2*phi-2*psi2true);
-	  float v2_track_reco = cos(2*phi-2*psi2reco);
+	  double v2_track_true = cos(2*phi-2*psi2true);
+	  double v2_track_reco = cos(2*phi-2*psi2reco);
 	  tp1d_v2pT_true->Fill(pt,v2_track_true);
 	  if(itrk>mult/2) tp1d_v2pT_reco->Fill(pt,v2_track_reco);
 	  tp1d_v2pT_pure->Fill(pt,v2);
 	  tp1d_YZ_v2pT_true->Fill(pt,v2_track_true*pt);
 	  tp1d_YZ_v2pT_reco->Fill(pt,v2_track_reco*pt);
 
-	  float x  = cos(phi);
-	  float y  = sin(phi);
-	  float xoff = offset;
-	  float yoff = offset;
-	  float xprime = x - xoff;
-	  float yprime = y - yoff;
-	  float phiprime = atan2(yprime,xprime);
+	  double x  = cos(phi);
+	  double y  = sin(phi);
+	  double xoff = offset;
+	  double yoff = offset;
+	  double xprime = x - xoff;
+	  double yprime = y - yoff;
+	  double phiprime = atan2(yprime,xprime);
 
-	  float v2_primetrack_reco = cos(2*phiprime-2*psi2reco);
+	  double v2_primetrack_reco = cos(2*phiprime-2*psi2reco);
 	  if(itrk>mult/2) tp1d_v2pT_prim->Fill(pt,v2_primetrack_reco);
 
-	  float v2_track_primereco = cos(2*phi-2*psi2primereco);
+	  double v2_track_primereco = cos(2*phi-2*psi2primereco);
 	  if(itrk>mult/2) tp1d_v2pT_primereco->Fill(pt,v2_track_primereco);
 
-	  float v2_track_doubleprimereco = cos(2*phiprime-2*psi2primereco);
+	  double v2_track_doubleprimereco = cos(2*phiprime-2*psi2primereco);
 	  if(itrk>mult/2) tp1d_v2pT_doubleprimereco->Fill(pt,v2_track_doubleprimereco);
 
-	  float u2x = cos(2*phi);
-	  float u2y = sin(2*phi);
+	  double u2x = cos(2*phi);
+	  double u2y = sin(2*phi);
 
-	  float twoprime = ( u2x*Q2x + u2y*Q2y - 1) / (mult - 1);
+	  double twoprime = ( u2x*Q2x + u2y*Q2y - 1) / (mult - 1);
 	  tp1d_d2pT->Fill(pt,twoprime);
 	  tp1d_YZ_d2pT->Fill(pt,twoprime*pt);
 	  tp1d_YZ_W->Fill(pt,pt);
 
-	  float u4x = cos(4*phi);
-	  float u4y = sin(4*phi);
+	  double u4x = cos(4*phi);
+	  double u4y = sin(4*phi);
 
-	  float fourprime = calc4_track(u2x,u2y,u4x,u4y,Q2x,Q2y,Q4x,Q4y,mult);
+	  double fourprime = calc4_track(u2x,u2y,u4x,u4y,Q2x,Q2y,Q4x,Q4y,mult);
 	  tp1d_p4pT->Fill(pt,fourprime);
 
-	  float fourprimeYZ = calc4_track_YZ(u2x,u2y,u4x,u4y,Q2x,Q2y,Q4x,Q4y,mult);
+	  double fourprimeYZ = calc4_track_YZ(u2x,u2y,u4x,u4y,Q2x,Q2y,Q4x,Q4y,mult);
 	  tp1d_YZ_p4pT->Fill(pt,fourprimeYZ*pt);
 
 	  ntracks++; // count total number of tracks
@@ -437,20 +437,20 @@ int main(int argc, char *argv[])
 
 
 
-float calc4_event(float Xn, float Yn, float X2n, float Y2n, float M)
+double calc4_event(double Xn, double Yn, double X2n, double Y2n, double M)
 {
 
-  float Qn2 = Xn*Xn+Yn*Yn;
-  float Qn2d = Xn*Xn-Yn*Yn;
+  double Qn2 = Xn*Xn+Yn*Yn;
+  double Qn2d = Xn*Xn-Yn*Yn;
 
-  float one   = Qn2*Qn2;
-  float two   = X2n*X2n+Y2n*Y2n;
-  float three = (2*(X2n*Qn2d + 2*Y2n*Xn*Yn));
-  float four  = 2*(2*(M-2)*Qn2);
-  float five  = 2*M*(M-3);
+  double one   = Qn2*Qn2;
+  double two   = X2n*X2n+Y2n*Y2n;
+  double three = (2*(X2n*Qn2d + 2*Y2n*Xn*Yn));
+  double four  = 2*(2*(M-2)*Qn2);
+  double five  = 2*M*(M-3);
 
-  float numerator = one + two - three - four + five;
-  float denominator = M*(M-1)*(M-2)*(M-3);
+  double numerator = one + two - three - four + five;
+  double denominator = M*(M-1)*(M-2)*(M-3);
 
   return numerator/denominator;
 
@@ -458,21 +458,21 @@ float calc4_event(float Xn, float Yn, float X2n, float Y2n, float M)
 
 
 
-float calc4_track(float xn, float yn, float x2n, float y2n, float Xn, float Yn, float X2n, float Y2n, float M)
+double calc4_track(double xn, double yn, double x2n, double y2n, double Xn, double Yn, double X2n, double Y2n, double M)
 {
 
-  float one   = (xn*Xn + yn*Yn)*(Xn*Xn + Yn*Yn);
-  float two   = x2n*Xn*Xn - x2n*Yn*Yn + 2*y2n*Xn*Yn;
-  float three = xn*Xn*X2n + xn*Yn*Y2n - yn*(X2n*Yn - Xn*Y2n);
-  float four  = 2*M*(xn*Xn + yn*Yn);
-  float five  = 2*(Xn*Xn + Yn*Yn);
-  float six   = 7*(xn*Xn + yn*Yn);
-  float seven = xn*Xn + yn*Yn;
-  float eight = x2n*X2n + y2n*Y2n;
-  float nine = 2*(xn*Xn + yn*Yn);
+  double one   = (xn*Xn + yn*Yn)*(Xn*Xn + Yn*Yn);
+  double two   = x2n*Xn*Xn - x2n*Yn*Yn + 2*y2n*Xn*Yn;
+  double three = xn*Xn*X2n + xn*Yn*Y2n - yn*(X2n*Yn - Xn*Y2n);
+  double four  = 2*M*(xn*Xn + yn*Yn);
+  double five  = 2*(Xn*Xn + Yn*Yn);
+  double six   = 7*(xn*Xn + yn*Yn);
+  double seven = xn*Xn + yn*Yn;
+  double eight = x2n*X2n + y2n*Y2n;
+  double nine = 2*(xn*Xn + yn*Yn);
 
-  float numerator = one - two - three - four - five + six - seven + eight + nine +2*M - 6;
-  float denominator = (M-1)*(M-2)*(M-3);
+  double numerator = one - two - three - four - five + six - seven + eight + nine +2*M - 6;
+  double denominator = (M-1)*(M-2)*(M-3);
 
   return numerator/denominator;
 
@@ -480,38 +480,38 @@ float calc4_track(float xn, float yn, float x2n, float y2n, float Xn, float Yn, 
 
 
 
-float calc4_event_YZ(float QTx, float QTy, float QT2x, float QT2y, float MQT)
+double calc4_event_YZ(double QTx, double QTy, double QT2x, double QT2y, double MQT)
 {
 
-  float one   = (QTx*QTx + QTy*QTy)*(QTx*QTx + QTy*QTy);
-  float two   = (QT2x*QT2x + QT2y*QT2y);
-  float three = 2*(QT2x*QTx*QTx + 2*QT2y*QTx*QTy - QT2x*QTy*QTy);
-  float four  = 2*(2*(MQT-2)*(QTx*QTx + QTy*QTy));
-  float five  = 2*MQT*(MQT-3);
+  double one   = (QTx*QTx + QTy*QTy)*(QTx*QTx + QTy*QTy);
+  double two   = (QT2x*QT2x + QT2y*QT2y);
+  double three = 2*(QT2x*QTx*QTx + 2*QT2y*QTx*QTy - QT2x*QTy*QTy);
+  double four  = 2*(2*(MQT-2)*(QTx*QTx + QTy*QTy));
+  double five  = 2*MQT*(MQT-3);
 
-  float cn4 = (one + two - three - four + five) / (MQT*(MQT-1)*(MQT-2)*(MQT-3));
+  double cn4 = (one + two - three - four + five) / (MQT*(MQT-1)*(MQT-2)*(MQT-3));
 
   return cn4;
 
 }
 
 
-float calc4_track_YZ(float pnx, float pny, float p2nx, float p2ny, float QTx, float QTy, float QT2x, float QT2y, float MQT)
+double calc4_track_YZ(double pnx, double pny, double p2nx, double p2ny, double QTx, double QTy, double QT2x, double QT2y, double MQT)
 {
 
   int mp = 1;
 
-  float one   = ((pnx*QTx + pny*QTy)*(QTx*QTx + QTy*QTy));
-  float two   = (p2nx*QTx*QTx - p2nx*QTy*QTy + 2*p2ny*QTx*QTy);
-  float three = (pnx*QTx*QT2x - pny*QTy*QT2x + pnx*QTy*QT2y + pny*QTx*QT2y);
-  float four  = 2*MQT*(pnx*QTx + pny*QTy);
-  float five  = 2*mp*(QTx*QTx + QTy*QTy);
-  float six   = 7*(pnx*QTx + pny*QTy);
-  float seven = (QTx*pnx + QTy*pny);
-  float eight = (p2nx*QT2x + p2ny*QT2y);
-  float nine  = 2*(pnx*QTx + pny*QTy);
+  double one   = ((pnx*QTx + pny*QTy)*(QTx*QTx + QTy*QTy));
+  double two   = (p2nx*QTx*QTx - p2nx*QTy*QTy + 2*p2ny*QTx*QTy);
+  double three = (pnx*QTx*QT2x - pny*QTy*QT2x + pnx*QTy*QT2y + pny*QTx*QT2y);
+  double four  = 2*MQT*(pnx*QTx + pny*QTy);
+  double five  = 2*mp*(QTx*QTx + QTy*QTy);
+  double six   = 7*(pnx*QTx + pny*QTy);
+  double seven = (QTx*pnx + QTy*pny);
+  double eight = (p2nx*QT2x + p2ny*QT2y);
+  double nine  = 2*(pnx*QTx + pny*QTy);
 
-  float dn4 = (one - two - three - four - five + six - seven + eight + nine + 2*mp*MQT - 6*mp) / (mp*(MQT-1)*(MQT-2)*(MQT-3));
+  double dn4 = (one - two - three - four - five + six - seven + eight + nine + 2*mp*MQT - 6*mp) / (mp*(MQT-1)*(MQT-2)*(MQT-3));
 
   return dn4;
